Sync use the BlockChain interface to check blocks
When bolt instantiates the Ethereum network layer, a in-memory
database is provided. That's because all structures persisted by
geth networking are temporary and don't make a lot of sense in our
persistence layer (Cassandra/Scylla). However the sync process
needs to be able to check the existence that was saved into our
persistence layer.

This changes the checks done by sync when getting block headers
from the beacon chain node to use the BlockChain interface,
backed by our persistence layer, instead of the default geth
database which in this case is in-memory only.