Use blockchain interface for canonical hash check in catalyst
The forkchoiceUpdated function was changed upstream to use rawdb.ReadCanonicalHash
for checking if a block is in the canonical chain. This doesn't work with custom
blockchain implementations (like prpl) that use a separate data store.

Reverted to using the blockchain interface method GetHeaderByNumber().Hash()
which works with any HandlerBlockchain implementation.